Added Ack plugin.  Added ,, mapping for C-^.  Changed ,ev and ,eg to not open in a split.
Fixed ,ev and ,eg to actually execute the :e command instead of leaving the cursor at the
command prompt.

Added Ack plugin and Command-T plugin leader key shortcuts.  In command mode %% expands
to the path of the current window's file.  This is useful for both Ack and Command-T.

For example with Ack:
",aSearchText %%"
This will search for "SearchText" using Ack from the current window's file's parent path.

And for Command-T it is used in the ,F mapping to search for files under the current
window's file's parent path.