1. Group autocmds with augroup and clear the group before adding the commands. This will ensure that autocmds don't get added more than once when resourcing the vimrc file.
2. I had broken my work around before checking it in by changing it to use
start to hide the command window that appears.  Turns out that doing this
either makes it not work, or makes it sometimes not work.  Sucking it up and
living with the annoying command window that appears when saving _vimrc or
_gvimrc.  I also changed these autocmds to work only on ~/_vimrc and ~/_gvimrc
instead of all files in ~, since it's only the rc files I expect to be symlinks.